third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.permission-self.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins in an iframe fail when the policy is 'self'
  assert_equals: expected "Failure" but got "Success"
Harness: the test ran to completion.
